Start working on ShapeEditor The ShapeEditor class should contain the logic used to manipulate Shapes in particular. Then, if/when other object types are implemented, they will get their own editor class. The ShapeEditor should be able to work pretty well, as it only has to track data that is pretty self-consistent: a series of selected points, and a small amount of data (e.g. rotation pivot) used to manipulate those points.
The ShapeEditor may become insufficient if/when shapes can support
multiple curves that utilize the same points - e.g. a beach ball
where all the arcs in the middle connect to the outer ring as part
of the same shape; this kind of thing will need to implemented,
though, because it would be vastly useful for animation.