Add 'correction grab'; make extrusion work better; misc The 'correction grab' mode might better be described as a 'line' grab mode. Basically, handles on endpoints of the selection are modified so that the curve points in straight lines on the edges of the selection. This basically makes it possible to move entire parts of a selection and also modify the curve around the selection so that it remains somewhat consistent.
Now, when an extrusion is made, correction grab is immediately set as
the editing state. This makes extrusions work better, because they leave
the shape in a somewhat consistent state.

Extrusions on endpoints of the curve, however, still work pretty
poorly--and this is complicated by whether a shape is looped / unlooped,
because I believe they will need to work differently based on that.

Other changes: when a split move is made, all points are deselected.
Perhaps in the future split moves will cause all points that were newly
added to be selected.