Begin working on some rearchitecting There's all sorts of cool stuff that I have planned for the "new" architecture.
First of all, it should support undoing and redoing -- and losslessly --
which is really cool, because I've sort of been dreading implementing
that stuff, but it is absolutely a necessary feature for this program to
be remotely usable. Incidentally, it also shouldn't be too difficult to
expand to a non-linear undo / redo, but that's probably for a later
date...

Second, each little "component" that is being edited is now going to be
responsbile for all of its "editing." (This is part of why undo / redo
will be easy). This will allow me to do some really cool stuff to get
reusable code. Namely, each "editor" class that wants to support a)
selections, and b) transforms will be able to implement an iterator that
hands the parent class each of its pieces of data, and then said parent
class will be the class where I implement selections and transforms.

This will also require an interface that the various editable data
supports as well, which will implement a) selection, and b)
transformations.

But overall it should make everything a whole lot simpler.